# openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
  title: RESTful EPP
  version: "1.0"
servers:
- url: http://localhost:8080/repp/v1/
paths:
  /:
    get:
      operationId: root_get
      parameters:
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Hello
      tags:
      - common
      x-openapi-router-controller: openapi_server.controllers.common_controller
...
  /messages:
    get:
      operationId: messages_get
      parameters:
      - description: Client transaction identifier
        explode: false
        in: header
        name: REPP-cltrid
        required: true
        schema:
          type: string
        style: simple
      - description: Namespace used
        explode: false
        in: header
        name: REPP-svcs
        required: true
        schema:
          type: string
        style: simple
      - description: Extension namespace used
        explode: false
        in: header
        name: REPP-svcs-ext
        required: false
        schema:
          type: string
        style: simple
      - description: Language used for response
        explode: false
        in: header
        name: Accept-Language
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/defaultOptionalRequestBody'
      responses:
        "200":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: OK
          headers:
            REPP-Queue-Size:
              $ref: '#/components/headers/queue-size'
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
        "224":
          content:
            application/epp+json:
              schema:
                type: string
            application/epp+xml:
              schema:
                type: string
          description: Unprocessable Content
          headers:
            REPP-cltrid:
              $ref: '#/components/headers/cltrid'
            REPP-eppcode:
              $ref: '#/components/headers/eppcode'
            REPP-svtrid:
              $ref: '#/components/headers/svtrid'
      summary: Poll request
      tags:
      - common
      x-openapi-router-controller: openapi_server.controllers.common_controller
...
components:
  headers:
...
    cltrid:
      description: Client transaction identifier
      explode: false
      required: false
      schema:
        type: string
      style: simple
    eppcode:
      description: EPP result code
      explode: false
      required: false
      schema:
        type: integer
      style: simple
    queue-size:
      description: Number of messages in the queue
      explode: false
      required: false
      schema:
        type: integer
      style: simple
    svtrid:
      description: Server transaction identifier
      explode: false
      required: false
      schema:
        type: string
      style: simple
...
  requestBodies:
    defaultOptionalRequestBody:
      content:
        application/epp+json:
          schema:
            type: string
        application/epp+xml:
          schema:
            type: string
      description: Default request body for XML of JSON message
      required: false
...